prevent auto-adding 'px' in css calc
In a css calc, it's enough that only one of the numbers has a unit.
e.g: calc(1px * 2)

Using the above example, in the previous behaviour,
incrementing '2' with up arrow will append 'px' to it: '3px'.
This CL avoid adding the unit when incrementing/decrementing
a number in a css calc value.

Bug:1126236
Change-Id: Ief4d9941a8d90ed763f4dc225307e2c49365bd5c