Typecheck timeline/TimelineUIUtils.js with TypeScript
This is a large change and unfortunately it is near impossible to
extract it in multiple parts. The primary reason for that is the
fact that the changes from objects to maps causes cascading changes
that require more refactoring. Sadly, you can't do one and not the
other, which means that the size creeps up.

In general, the changes fall in the following buckets:
- Use module variables instead of static properties
- Use maps instead of objects
- Refactoring AggregatedStats to separate setting the `idle` field
- Extract binaryIndexOf (only usage was in this file)

R=jacktfranklin@chromium.org

Bug: 1011811
Change-Id: Ice1107738e34b0a926985aa657b46f34201f0d3b