Improve _searchHighlighting performance
This change advances the stream by manually setting the stream position to the next occurrence of the matched pattern.

Before this fix, the stream advanced by calling the codemirror api, stream.next(), which advances position by 1 every call. For longer patterns, this causes the debugger to slow down significantly or even freeze.

Fixed: 1127708
Change-Id: I5572b79216c8231d4c5b9148b991307684f1b4d5