Allow shorthand CSS properties to be displayed when writing longhand CSS properties in uppercase letters.
This bug was caused when the user passes in longhand CSS properties written in uppercase letters.  The browser would not display 'toggle-expandable' triangle indicating the availability of shorthand CSS properties.

The fix was to add toLowerCase() to the name argument inside of the longhandsProperties() method.  This allows the argument to be passed into the longhands() method as a lowercase string, which then is able to match the lowercase string and returns an array of shorthand CSS properties.

Fixed: 918109
Change-Id: I018b3e7eccaa008ec2b9dbff101bb5681655e624