[tests] Don't print steps except for error cases
console.log() output is not buffered during parallel testing so it's
hard to associate the step() logging output with the test that
actually caused it. If we want to log stuff like this, we should find
a way to redirect it to the test reporter properly so that it gets
buffered per-test-file and printed all at once, like the describe/
it output is.

The step output probably isn't needed anymore as we have the AsyncScope
output which gives the source-mapped file name + line number in the
test output if there is an await failure, which is even more useful
than the step output.

If people find cases where we do really need this we can add it back,
the step function still exists for now.

Bug: 1101784
Change-Id: Ia90b6532ce348b189ee7c5f2e3ed482f074a6561