Enable NodeText colours to be configured by parent
As part of the new dark theme, we want to change some of the colours
used for the ElementsBreadcrumbs, including those where it shows an
element's ID and class.

We long term probably want to have the same colours for every use of
this component, but to enable us to push the new theme colours into the
breadcrumbs without impacting other uses, this CL adds CSS variables
that the parent component can define, which the NodeText component will
then use.

Bug: 1148733
Change-Id: I0c8d26dd661cbb3add10c7cc98062d239a9bac2a