[tests] Run chromium before prompting during interactive e2e mode
With the prompting in the global setup fixture, we started the hosted
mode server and then paused with the user prompt. This is not how it
worked before, where we launched chrome before we paused.

Set it back to the previous behavior by moving the prompt code inside
the beforeAll hook. Debug mode is mututally exclusive with parallel
mode so the moved code will only run in serial mode, not once per
parallel process.

Bug: 1101784
Change-Id: I74c9a70652056ecf368e439ee747409742ff4ce9