Change ::selection backgrounds according to Blink fix
Blink did a fix to the ::selection background painting to only do
a transformation to add transparency for replaced content. This was a
fix to align with the specification and the Gecko implementation.

This caused a regression in the devtools front-end ui relying on that
transformation. Change the colors to end up at the same end result as if
the transformation was still done. This will still work for the old
Blink implementation because no transformation is done if there is an
alpha component < 1.

Note there are various ::selection colors in the devtools stylesheets
which already had an alpha of .99 in order to work around that bug.

Bug: 1149259, 1018450
Change-Id: I8d42cd7f73c35003793eb363385be499c0e8aae9