Make Linkifier actions take LinkInfo objects
In order to enable a new web component to emit events to be caught by the
Linkifier and processed, we need to first change the structure slightly such
that the link processsing functions take the `LinkInfo` object, not the node
itself. This will enable a new component to emit a LinkInfo object that can be
processed (in follow up CLs). See the bug attached for more details.

Bug: 1149403
Change-Id: Ia4d04192c13f33c85368811f5cffe9931148bcd4