[wasm] Add informative console messages when loading debug symbols.
This adds some console log messages shown while loading debug symbols
via a debugger language plugin (aka a language extension point), which
should make it easier to tell what's going on and when the debugger
is expected to be in "ready" state.

Turn `DebuggerLanguagePlugin` into a base class instead of an interface
(we can later add a proper TypeScript interface here that is also shared
with the DWARF extension), and introduce a `name` property which yields
the user visible name of the extension that we can use in log messages.

After: https://imgur.com/NoOV3zy.png
Bug: chromium:1083146, chromium:1126328, chromium:1136484
Change-Id: I9d7a5b7db80964445911bb28b8ac423e0f303b70