[unit test] Adds missing entrypoints
We don't currently import all entrypoints for testing, which in turn
means that our coverage report is incomplete.

This CL adds the remaining imports such that we get a complete coverage
report. In time the plan is to move to importing
front_end/devtools_app.js as the canonical, single entrypoint, and avoid
the need to list out all entrypoints here.

R=tvanderlippe@chromium.org

Bug: 1148296
Change-Id: Iffa183d36b764d4465f20c4e38227f930f188d62