[sdk][bindings] Introduce DebuggerModel._expandCallFramesCallback.
This adds a `DebuggerModel._expandCallFramesCallback` which is an async
callback that's invoked when the `DebuggerModel` receives a paused event
and before it propagates this information to its observers. The callback
is provided by the `DebuggerLanguagePluginManager`, which utilizes its
plugins to expand the call frames in the presence of inlining.

This breaks one of the dependency cycles where the DebuggerModel
depends on the bindings/ module.

Bug: chromium:1150029, chromium:1083146, chromium:1126943
Change-Id: I66908bad4a01086483e400b2cdb31fb0355c4b99