[sdk][bindings] Introduce DebuggerModel._computeAutoStepRangesCallback.
This adds a `DebuggerModel._computeAutoStepRangesCallback`, an async
callback, which is invoked by the `DebuggerModel` when stepping is
requested to compute the location ranges that should be skipped. The
callback is provided by the `DebuggerWorkspaceBinding`, which utilizes
the language plugins (if any) to compute the correct step range (also in
the presence of inlining). This is currently limited to language
plugins, but can and should be extended to source maps in the future.

This breaks one of the dependency cycles where the DebuggerModel
depends on the bindings/ module.

Bug: 1150029, 1083146, 1136814, 1137302, 1018234
Change-Id: I524f12d579c98c18b2feb9a7510f917fd9bb0f4a