Tweak metrics pane styles
Currently the titles and values for the padding metrics can overlap if
the styles of the target element have a large value for the vertical
padding, and a low amount for the horizontal padding, e.g. 120px 0;

This CL positions the labels for margin, border, and padding slightly
further to the left to create more horizontal space, and sets a minimum
width for the padding box so that it's less likely to overlap.

Fixes: 1149935
Change-Id: Ie6f217c7346955c03f93033cd42e19e5c28cb901