[perf] Restore boot perf test
We updated the e2e tests to using a different architecture a while ago
but we didn't update the boot perf tests to match. This CL migrates
them to using node_ts_library in the same way as e2e, sets the output
directory to be perf-data in the root of front_end, and ensures that
the JSON file contains mean and percentile data.

Change-Id: I0738ae664e52ec7a8fa76173a0155e14f583259f