Adjust handling of variable-2 tokens so they are treated as properties
This change makes it so custom CSS properties will be formatted as
properties. Custom CSS properties are being labeled as "variable-2" type
and there was no handling in the _tokenCallback located in CSSFormatter.js
for these token types.

An additional regex check for "variable-2" token types in the if
statementment which manages "property" token types to allow for the same
treatment.


Fixed: 1024752
Change-Id: Ib47b3113755eed4f8962fbec1a7e4b8c43615c4e