Fixes revealing of the Settings screen by redirecting
The introduction of an interface to group legacy and new Setting types for extensions migration affected the revealing of the settings screen when redirected from other parts of the frontend. This because this revealing relies on a lookup based on the 'instanceof' keyword, which fails because of the new usage of interfaces. This fixes the issue by using a super class instead of an interface to group the Setting types.

Bug: 1150708, 1134103
Change-Id: I92d067d955808bb9de4dcaa4a635dd58380934b1