Screenshot: disable screenshot command in PWA
Previously, users could use the command prompt menu to select
screenshot tools such as "take area screenshot" and "take full
"screenshot".

However, in modes where the devtools cannot be
docked, such as PWA, the device mode that handles the
screenshot is never created, so the screenshot actions don't do
anything.

This CL disables the screenshot commands using the same "can_dock"
check as the emulation mode, so that screenshots cannot be attempted
in this non-supported case.

Before:
After:
Change-Id: I758f1ca3c0d3b4ad9575e6a4a1a38e0919d45584