[wasm] Introduce unregisterLanguageExtensionPlugin API.
This adds a new method `unregisterLanguageExtensionPlugin()` to the
`LanguageServicesAPI` exposed to DevTools extensions, which allows the
developer to unregister a previously registred language extension
plugin. This is useful for example when the configuration for an
extension changes and the front-end should reload the source mappings.

Drive-by-fix: Also make the registerLanguageExtensionPlugin and the
unregisterLanguageExtensionPlugin methods return promises, so that
the extensions can properly await / chain them.

Fixed: chromium:1151285
Bug: chromium:1151280, chromium:1041362, chromium:1083146
Change-Id: Ib62b70c717754d70cf827ba1a97fb5c3e358ee1f