Update NodeText to apply classes for multiple descriptors
One of the things we want to do with the NodeText on breadcrumbs is
apply different colours to the ID/classes in the breadcrumb depending on
if the Node has just one "descriptor" (an ID or a class) or multiple (ID
and class(es)). Long term when we've completed the dark mode work we
won't need these extra variables, we'll just define them in the NodeText
class itself, but for now because we want to migrate piece by piece we
need to apply these classes and accept extra CSS vars to enable us to
colour the breadcrumbs properly for the new dark (and light) mode work.

R=aerotwist@chromium.org

Bug: 1122511
Change-Id: I1c510261c2a81f668435399a674bbf45e8ae22ca