Flex: Draw self-alignment at the container level in flex overlay
This draws a thicker line where the align-items value of the flex
container is, as well as an arrow (or 2 arrows) pointing to where
alignment happens.

Baseline alignment is missing as it will require specific support from
the backend (the offset of the baseline).

Backend CL: https://chromium-review.googlesource.com/c/chromium/src/+/2552282/

Screenshots: https://imgur.com/a/uvkKJIl

Bug: 1139949
Change-Id: Ie9387714abc8b9953deb38e5b89f5346206dcc13