Fix Node inspect profiler
While js_profiler is an empty module, it has a single purpose:
specify a panel in its `module.json`. Therefore, it needs a corresponding
(unused) entrypoint file, with the appropriate GN infrastructure.

While the files itself are basically unused, the Runtime enforces
that every single module has a corresponding entrypoint. And that
assumption was invalid for `js_profiler`.

R=aerotwist@chromium.org

Fixed: 1151855
Change-Id: I14acce9c547c22c67a3b4bb26e8a7a2224015254