[refactor] Align `LocationRange` and `Location` classes more closely.
The `SDK.DebuggerModel.LocationRange` is now the range equivalent to the
`SDK.DebuggerModel.Location` class, and can be widely used throughout
the DevTools front-end - instead of having ad-hoc pairs of start and end
`Location`s. The first step here is to use `LocationRange` consistently
across the language plugins. The eventual vision here is to switch more
places from `Location` to `LocationRange` to preserve the range
information, and eventually have source maps also report in terms of
ranges instead of points (and thus align source map DX with the language
plugin DX).

The sorting and merging for the skip lists that we send with the
stepping commands is now done on the protocol level structures, which
also simplifies the testing a bit.

Bug: chromium:1083146, chromium:1018234, chromium:1105765
Change-Id: Ia400eeff01a59ecbf15b86f9dddecd9e3ee2f174