[DataGrid] support filter parser input
`TextUtils.FilterParser` is commonly used to parse user input into a
series of filters. This CL updates the new data-grid-controller so it
can take the result of parsing user input into an array of parsed
filters. This means we can hook it up to existing input boxes in
DevTools and unblocks us landing the new data grid into the protocol
monitor.

Bug: 1125968
Change-Id: I45896796d7e7cc720d22c48b65325fb099115da8