Do not await runIfWaitingForDebugger command
For websites that are using rel="noopener" links (which became
the default in M89), DevTools would halt the debugger when being
opened. The culprit was an `await` added to the `runIfWaitingForDeugger`
command during typescriptification. Reverting to the previous
"fire-and-forget" logic by avoiding the `await` restores the correct
behavior for rel="noopener" links.

R=aerotwist@chromium.org

Also-By: ericlaw@microsoft.com
Fixed: 1152727
Change-Id: I25caf698e4ed64e98a08997923fb1f459df493b3