[ESLint] rule for .data=${{} as X}
This rule enforces that within a LitHtml template when we bind to the
`.data=` property we provide an explicit type cast that must be a type
reference.

This is one of the checks that the component bridges used to make, but
now we need to port them to ESLint rules.

Bug: 1130536
Change-Id: I401661289a829fd181bf91943fffe162b87de8b8