[debugger] Don't ask V8 for breakpoints in case of language plugins.
When the source code is provided by language plugins, they should also
report appropriate break locations, and don't rely on the break
locations reported by V8. In particular, in case of WebAssembly, the
WebAssembly engine basically reports all Wasm instructions, which is
not particularly useful.

Note that we need a similar change for source maps.

Also-By: pfaffe@chromium.org
Bug: chromium:1153123
Change-Id: I5559446e607fa26f42c07b3ec5b090f4bbc05a5a