Grid: Bring back some simple luminance calculation for label text color
We used to import the whole Color module to do this in the past.
Then we decided to remove it as it was making the bundle too big. Plus
all of the default label colors worked great with the default #121212
text color, so we just hard-coded that text color.

This brings back some simple, light-weight, luminance calculation that
we use to decide between #121212 and white text color.

The calculation does retunr the usual #121212 color for all of our
default label colors, so there's no changes there. But it returns white
if the label color is changed by the user and becomes too dark.

Bug: 1153292
Change-Id: I45cba8cd0033e8be77e60c7392e2415bddc7599f