Remove twin definitions from ConsoleModel
ConsoleModel was defining twin enums for protocol definitions. This
was a brittle approach, because those enums needed to be kept in sync
with the CDP enums by hand. Additionally, we didn't get good
type-checking, as the protocol types had to be up-cast to string to
match the ConsoleModel's interface.

This CL uses proper types to enhance type-checking, and strengthens
the ConsoleModel's interface types accordingly.

We need the .ts file to get the right type for the enums.

Bug: chromium:1153099
Change-Id: I9b669a65531bad560a00c6feb4ea4c6cd29679e1