Fix running layout tests
In crrev.com/c/2520088 the blink web_tests infrastructure started using
a module that is in Python 3, but not 2. This started crashing our `npm
run test`, since we force the python version. Based on the comments on
the original CL, we should call the Python script directly and then
everything works as expected.

R=aerotwist@chromium.org

Change-Id: Iafe123e6adf792da940d3ca605e2eddd49ee5eba