[DarkMode] Use `color-link` for links
Replace the legacy `link-color` with the new `color-link` for the color
of links. In dark mode links will now be blue instead of grey.

This change only affects links which were using `link-color` to set
their color. This includes the widely used XLink element. Notable
exceptions where the link color is not being changed are the Elements
Panel and the Console.

Bug: 1152736

Change-Id: If938f4acd59133327d5711877b17b79b2df93bd2