Flex: Support for highlighting align-items:baseline in the flex overlay
A container can define align-items:baseline to get its items to all
align according to their respective baselines.
Because of this, this alignment value does not lead to an absolute
position within the container like flex-start/end/center/stretch do
(where, e.g., flex-start is always at the start edge of the container's
cross edge).
Instead, the alignment position depends on all of the item's font and
font-size.

To know where the common baseline used to align items is, the only
information at our disposal is each item's baseline value.
(see backend CL: https://chromium-review.googlesource.com/c/chromium/src/+/2567643/)

In theory though, items can override their own self-alignment, making it
impossible for us to know if a given item's baseline is actually the one
that some items are aligned on.
I think reality is otherwise, in the vast majority of cases. I think
that when a container defines baseline alignment, most if not all items
are actually aligned on their common baseline. So it is safe, in most
cases, to just use the item's baselines to draw where alignment happened
in the overlay.

Screenshot: https://imgur.com/oN0plAf.png

Bug: 1139949
Change-Id: I714a63e649a15d3c5f963399a2bdf6ff019371be