[DataGrid] add dark mode colors
This CL updates the new data-grid to work in dark mode.

There's also a drive-by fix for a bug where if the user clicked the cell
that the code had deemed as focusable, the row wouldn't be highlighted
as selected. We were trying to avoid re-rendering if the user clicked
the cell that is already focused, but in actual fact because they've
clicked the cell marked as focused, we need to re-render so that the
right row gets highlighted as selected. Unfortunately I can't test this
change in Karma with unit tests because I need to "properly" click and
achieve focus, but I will add a test once interaction tests are
possible.

Fixed: 1153064
Change-Id: I471c707545aab03535d13ea2a203752b897dafa1