[Loc] Fail presubmit on modifications to en-US.json
Currently presubmit doesn't fail when it modifies en-US.json. This
allows for the possibility of uploading patches with en-US.json out of
sync with strings found in the product.

This patch forces presubmit to fail if it modifies en-US.json, forcing
the developer to check-in the correct json file.

Only en-US.json is checked since its CTC and XL counterparts aren't
managed by version control.

DISABLE_THIRD_PARTY_CHECK = This patch makes some logging in presubmit
redundant so it needs to be removed.

Bug: 1136655
Change-Id: Ib2b78bb4fe3cb688de915cb6ffa2868e78a86c07