Service Worker Update Cycle Details
This CL adds SW update cycle details to the Service Workers view,
Application pane. See the explainers below for the requirements.

A summary of features implemented:
- For the service worker associated with the current target, if
it has been installed and/or activated, show a timeline of the
installation and/or activation events.
- Such timeline is based on the observation in the frontend and
is intended to show for example, if an event takes extra long
time. It is not meant to be a measurement of execution time of
the event handler at the backend.
- Each event's timeline may be expanded to reveal details of
the event handler, such as the starting and ending time of
the event handling, as well as cache entries being added or
deleted.
- The requirement in the explainer for the promise associated
with the installation/activation is dropped as it is not clear
such a promise adds much value.

crbug.com/1066604

Explainer:
https://docs.google.com/document/d/1M2xKrlFHrojGHIJ9AVpQdGrmUUU0G8YVToa3kZz3RZA/edit?usp=sharing

Change-Id: I0b3725b342eeb32470ef759bab8a7da5a350dba9