[debugger] Don't load sources from stale raw modules.
When addRawModule returns we only check whether the DebuggerModel is
still there, but we don't check that the script is still valid (in case
of the main thread, the DebuggerModel remains the same, only the scripts
disappear on reload). So this change adds the missing check here before
reporting the sources files.

Also we shouldn't log the completion message upon responses to stale
requests, as that's really confusing to the developer.

Also-By: pfaffe@chromium.org
Bug: chromium:1155067, chromium:1083146
Change-Id: I8a9fd2a8413d4ac1a543b3bf802ad5892fa091c9