Add support for hashbang comments when formatting JavaScript
Acorn has a separate option to turn on compatibility for hashbang
comments. We need to turn on this flag for both our parser and
tokenizer.

Sadly, this wasn't enough to make the formatting work. What we discovered
is that comments can arrive out-of-order. In the case of hashbang
comments, we first receive the comment and then the token. However,
for all other comments, we first receive the token and then the comment.

To combat this reordering problem, we have to update our constructor
to only perform initializations when we don't have any comments at the start
(which is the situation for regular comments). Then when we do have
a hashbang comments, we bail out of shifting the comment away (and thus
would lose track of it) and let both `peekToken` and `nextToken` process
the comment.

R=bmeurer@chromium.org

Also-By: jacktfranklin@chromium.org
Fixed: 942422
Change-Id: Ic1971072b9c2d1e3ffeda7c3f9eede3b10f669fe