Allows the pretty printing of directly visited minified asset files.
PreviewFactory.js
const highlighterType is being set as 1 of 2 variables but the order in which it checked these variables was incorrect. The current mimeType should be checked prior to the canonicalMimeType9(). Reversed the order.

FormatterWorker.js
Directly visited minified asset files are given a mimeType of application/javascript on page load. On our FormatterWorker file, we did not have a case that would format this mimeType. Added the mimeType of ‘application/javascript’ to our switch statement allowing it to format with the same settings currently set on mimeType ‘text/javascript’.

R=jacktfranklin@chromium.org

Bug: 1019785
Change-Id: Ib43e9c255129da95492bdfeaba7651964685626f